project/generation/webapp/app/requirements.txt
click==8.1.3 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
faker==18.3.1 ; python_version >= "3.9" and python_version < "4.0"
flask==2.2.3 ; python_version >= "3.9" and python_version < "4.0"
importlib-metadata==6.1.0 ; python_version >= "3.9" and python_version < "3.10"
itsdangerous==2.1.2 ; python_version >= "3.9" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.9" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.9" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
mypy==1.1.1 ; python_version >= "3.9" and python_version < "4.0"
mysql-connector-python==8.0.32 ; python_version >= "3.9" and python_version < "4.0"
protobuf==3.20.3 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.9" and python_version < "3.11"
typing-extensions==4.5.0 ; python_version >= "3.9" and python_version < "4.0"
werkzeug==2.2.3 ; python_version >= "3.9" and python_version < "4.0"
zipp==3.15.0 ; python_version >= "3.9" and python_version < "3.10"
